automl/requirements.txt
nltk
matplotlib
pandas
wordcloud
google-cloud-aiplatform
kfp
google-cloud-pipeline-components